fix(tf): Use folder_id/org_id for project parent
The 'google_project' resource expects 'folder_id' or 'org_id'
instead of the generic 'parent' argument.

This commit corrects the 'iac/terraform/modules/project/main.tf'
file to use the correct argument, and adds logic to parse the
parent variable to determine whether it is a folder or an
organization.