feat(cuj02): Add org-less single-user setup
This commit introduces a new, cost-effective single-user setup (CUJ02)
that supports the creation of org-less projects.

This new setup is triggered by a `just user-up` command and is configured
with a YAML file. It uses the existing `1b_single_user_setup` Terraform
module, but with a simplified configuration that does not require a parent
folder or organization.

This resolves issue #22.